refactor(DC-704): extract Serilog/OTLP logging setup into LoggingSetup
